Only show active titles
- and simplyfied process

+++ app/src/main/java/com/ceaver/assin/markets/MarketCompleteUpdateWorker.kt
@@ -11,18 +11,21 @@ class MarketCompleteUpdateWorker(appContext: Context, workerParams: WorkerParame
         val allRemoteTitles = MarketRepository.loadAllTitles()
         val allLocalTitles = TitleRepository.loadAllTitles()
 
-        val newTitlesToInsert = allRemoteTitles - allLocalTitles
-        val existingTitlesToUpdate = (allRemoteTitles - newTitlesToInsert).map { if (it.active.toInt() > 99) it else it.copy(active = Integer(it.active.toInt() + 1)) }
-        val removedTitles = allLocalTitles - allRemoteTitles
-        val removedTitlesToDelete = removedTitles.filter { it.active < -99 }
-        val removedTitlesToUpdate = (removedTitles - removedTitlesToDelete).map { it.copy(active = Integer(it.active.toInt() - 1)) }
+        if (allLocalTitles.isEmpty()) { // initial load
+            TitleRepository.updateAll(allRemoteTitles.map { it.copy(active = Integer(50)) }.toSet())
+        } else {
+            val newTitlesToInsert = allRemoteTitles - allLocalTitles
+            val existingTitlesToUpdate = (allRemoteTitles - newTitlesToInsert).map { it.incrementActiveCounter() }
+            val removedTitles = allLocalTitles - allRemoteTitles
+            val removedTitlesToDelete = removedTitles.filter { it.inactive() }
+            val removedTitlesToUpdate = (removedTitles - removedTitlesToDelete).map { it.decreaseActiveCounter() }
 
-        TitleRepository.updateAll(newTitlesToInsert + existingTitlesToUpdate + removedTitlesToUpdate)
-        TitleRepository.deleteTitles(removedTitlesToDelete.toSet())
+            TitleRepository.updateAll(newTitlesToInsert + existingTitlesToUpdate + removedTitlesToUpdate)
+            TitleRepository.deleteTitles(removedTitlesToDelete.toSet())
 
-        newTitlesToInsert.forEach { LogRepository.insertLog("Added  ${it.name} (${it.symbol}) to local database.") }
-        removedTitlesToUpdate.forEach { LogRepository.insertLog("Decreased active rating of ${it.name} (${it.symbol}) (${it.active})") }
-        removedTitlesToDelete.forEach { LogRepository.insertLog("Removed  ${it.name} (${it.symbol}) from local database.") }
+            existingTitlesToUpdate.filter { it.active.toInt() == 50 }.forEach { LogRepository.insertLog("Activated  ${it.name} (${it.symbol}).") }
+            removedTitlesToDelete.forEach { LogRepository.insertLog("Removed  ${it.name} (${it.symbol}) from local database due to long time inactivity.") }
+        }
 
         return Result.success()
     }
